Fix unsafe buckets turning into named safe buckets.
Unsafe buckets are distinguished by a custom "GLOW" enchant, as well as
having a custom name.  In order for a bucket to dispense unsafe water, it
must have the custom enchantment.

The GLOW enchantment does not exist immediately after the server starts.
It is lazily created the first time someone asks for an unsafe water bucket
with "/sb unsafe".  If nobody runs that command after a server restart,
ModMode will be unable to correctly deserialize unsafe buckets from a
moderator's ModMode inventory; they will lack the GLOW enchant.

The fix is simple: force creation of the custom GLOW enchant when the
SafeBuckets plugin is enabled.