Fix dispenser handling and improve documentation.
Prior to this commit, in order to make a dispenser dispense flowing liquid
you needed to left click on the dispenser with the tool item and also left
click the output side of the dispenser with the tool block.  The latter
action caused the location of the dispensed liquid to be marked as unsafe
(flowing).  The process was needlessly complicated and admins had commented
on their confusion in getting it to work.

The new code simply sets the dispensed liquid as flowing if the dispenser
block has been marked as unsafe.  So, you only have to left click on the
dispenser with the inspector tool.

The behaviour of the inspector tool item and block has been clarified in the
documentation, and the use of these tools with dispensers is explained.

The code for handling the inspector tools has been refactored for brevity and
slightly better performance, and colour added to inspector tool messsages.