Fix long-standing adjacent-to-waterlogged-always-flows bug
The bug could be replicated by placing a waterloggable block (e.g. slab), dumping a bucket of water on the slab to make it waterlogged, and then dumping another bucket of water on an adjacent block by clicking on one of the slab's block faces. The newly-placed adjacent water block will then flow when it should not.